requirements.txt
aiocache==0.11.1
aiofiles==0.8.0
aiohttp==3.8.3
alembic==1.9.4
amis-python==1.0.6
anyio==3.6.2
APScheduler==3.9.1.post1
arrow==1.2.3
asyncio==3.4.3
attrs==22.1.0
bbcode==1.1.0
bs4==0.0.1
cachetools==5.2.0
click==8.1.3
cn2an==0.5.17
dateparser==1.1.1
easyocr==1.6.2
EdgeGPT==0.1.22
emoji==2.1.0
expandvars==0.9.0
expiringdict==1.2.2
Faker==16.4.0
fastapi==0.88.0
feedparser==6.0.10
fuzzywuzzy==0.18.0
githubkit==0.9.6
GitPython==3.1.29
h11==0.12.0
h2==4.1.0
httpx==0.23.0
humanize==4.4.0
idna==3.4
importlib-metadata==4.12.0
iso8601==1.1.0
jieba==0.42.1
Jinja2==3.1.2
langdetect==1.0.9
littlepaimon-utils==1.0.2
loguru==0.6.0
lxml==4.9.1
Mako==1.2.4
matplotlib==3.6.2
nb-cli==0.6.7
nonebot-adapter-onebot==2.2.0
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-chatrecorder==0.2.3
nonebot-plugin-datastore==0.6.2
nonebot-plugin-gocqhttp==0.6.10
nonebot-plugin-guild-patch==0.2.2
nonebot-plugin-htmlrender==0.2.0.1
nonebot-plugin-imageutils==0.1.13.4
nonebot-plugin-localstore==0.2.0
nonebot2==2.0.0rc2
numpy==1.23.4
openai==0.27.0
paho-mqtt==1.6.1
parsel==1.7.0
peewee==3.15.3
pil-utils==0.1.6
playwright==1.27.1
psutil==5.9.4
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
pypinyin==0.47.1
python-jose==3.3.0
python-socketio==5.7.2
pytz==2022.6
qrcode==7.3.1
regex==2022.3.2
requests==2.28.1
rply==0.7.8
rtoml==0.9.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
SQLAlchemy==2.0.9
tinydb==4.7.0
tortoise-orm==0.19.2
tqdm==4.64.1
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
win32-setctime==1.1.0
wordcloud==1.8.2.2